made clear_database an option
