+++ crate/server/src/config/db.rs
@@ -31,6 +31,11 @@ pub struct DBConfig {
     /// The directory path of the sqlite or sqlite-enc
     #[clap(long, env = "KMS_SQLITE_PATH", default_value = "./sqlite-data")]
     pub sqlite_path: PathBuf,
+
+    /// Clear the database on restart
+    /// WARNING: This will delete all the data in the database
+    #[clap(long, env = "KMS_CLEAR_DATABASE", default_value = "false")]
+    pub clear_database: bool,
 }
 
 impl Default for DBConfig {
@@ -39,6 +44,7 @@ impl Default for DBConfig {
             database_type: "sqlite".to_string(),
             database_url: None,
             sqlite_path: PathBuf::from("./sqlite-data"),
+            clear_database: false,
         }
     }
 }
