+++ crate/server/src/tests/test_utils.rs
@@ -36,6 +36,7 @@ pub fn https_clap_config() -> ClapConfig {
             database_type: "sqlite".to_string(),
             database_url: None,
             sqlite_path,
+            clear_database: true,
         },
         ..Default::default()
     }
